V3- apiConnect

See merge request 2024-2025-but-info2-a-sae34/l1/l11/but-info2-a-sae3-docker-stack!8

+++ sfapp/src/Controller/BuildingController.php
@@ -31,7 +31,7 @@ class BuildingController extends AbstractController
             return $this->redirectToRoute('app_error_403');
         }
         return $this->render('building/index.html.twig', [
-            'buildings' => $buildingRepository->findAll(),
+            'buildings' => $buildingRepository->sortBuildings(),
         ]);
     }
 
@@ -63,10 +63,12 @@ class BuildingController extends AbstractController
     }
 
     
-    #[IsGranted("ROLE_ADMIN")]
     #[Route('/building/{id}', name: 'app_building_delete', methods: ['POST'])]
     public function delete(Building $building, EntityManagerInterface $entityManager): Response
     {
+        if (!$this->isGranted('ROLE_ADMIN')) {
+            return $this->redirectToRoute('app_error_403');
+        }
         $entityManager->remove($building);
         $entityManager->flush();
